Add files via upload
hi, I have restarted the home page with a bit of the glossary information. The HTML is further along than the CSS I have added some placeholder images so you can use them if you like in the comments it says which one goes where. The CSS is pretty bare bones but I was just experimenting up to this point. On Monday evening I will be carrying on with this in a more focussed way so feel free to branch and change things before then.

V